+++ src/main/java/br/ufs/gothings/plugins/coap/CaliforniumServer.java
@@ -75,7 +75,7 @@ public class CaliforniumServer {
             final GwHeaders h = gw_request.headers();
             h.set(GW_PATH, uri);
             final CoAP.Type type = exchange.getRequest().getType();
-            h.set(GW_QOS, (byte) (type == CoAP.Type.CON ? 1 : 0));
+            h.set(GW_QOS, type == CoAP.Type.CON ? 1 : 0);
 
             // Set headers according to the request code
             switch (code) {
